backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Computer Vision
opencv-python-headless==4.9.0.80
PaddleOCR==2.7.3
PaddlePaddle==2.6.1
numpy==1.26.4
Pillow==10.2.0

# AI / ML
torch==2.2.1
transformers==4.39.1

# RAG
faiss-cpu==1.8.0
sentence-transformers==2.5.1

# API clients
httpx==0.26.0

# Utils
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
